@@ -1,9 +1,8 @@
 import React from "react";
-import { Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { routes } from "./routes";
 import { AuthRoute } from "./feature/Common/AuthRoute";
 import { Layout } from "./feature/Common/Layout";
-import { history } from "./helpers/history";
 import { Provider } from "react-redux";
 import { store } from "./store";
 
@@ -22,29 +21,34 @@ initAuthInterceptor(store, logout);
 const App = () => {
   return (
     <Provider store={store}>
-      <Router history={history}>
-        <Layout>
-          <Switch>
-            {routes.map((item, i) =>
-              item.isAuth ? (
-                <AuthRoute
-                  path={item.path}
-                  component={item.component}
-                  key={i}
-                  exact={item.exact}
-                />
-              ) : (
-                <Route
-                  path={item.path}
-                  component={item.component}
-                  key={i}
-                  exact={item.exact}
-                />
-              )
-            )}
-          </Switch>
-        </Layout>
-      </Router>
+      <Layout>
+      <BrowserRouter >
+          <Routes>
+            {routes.map((item, i) =>{
+              const El = item.component
+              return(
+                item.isAuth ? (
+                  <Route
+                    path={item.path}
+                    element={
+                      <AuthRoute>
+                        <El/>
+                      </AuthRoute>
+                    }
+                    key={i}
+                  />
+                ) : (
+                  <Route
+                    path={item.path}
+                    element={<El/>}
+                    key={i}
+                  />
+                )
+              )}
+                )}
+          </Routes>
+      </BrowserRouter>
+      </Layout>
     </Provider>
   );
 };
