@@ -1,7 +1,8 @@
 import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { withRouter, Link } from "react-router-dom";
+import { Link } from "react-router-dom";
+import {withRouter} from './../../../../helpers/withRouter'
 import { confirmRestorePasswordRequest } from "../actions";
 import { Loader } from "../../../Common/Loader";
 import style from "./index.module.scss";
