@@ -1,9 +1,10 @@
 import React, { useEffect } from "react";
 import PropTypes from "prop-types";
-import { SignupForm } from "../components/SignupForm";
+import SignupForm  from "../components/SignupForm";
 import { connect } from "react-redux";
 import { signupRequest } from "../actions";
-import { withRouter, Link } from "react-router-dom";
+import { Link } from "react-router-dom";
+import {withRouter} from './../../../../helpers/withRouter'
 import style from "./index.module.scss";
 
 const Signup = props => {
