@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { withRouter } from "react-router-dom";
+import { withRouter } from "./../../../../helpers/withRouter";
 import { usersFetchRequest } from "../actions";
 import { Loader } from "../../../Common/Loader";
 import { Table } from "antd";
